requirements.txt
pandas
seaborn
numpy
matplotlib
scikit-learn
catboost
xgboost
dill
Flask
DVC
pyyaml
# -e . // Keep in comment when you install requirements we will do uncomment when we want to build it.